boot/nxboot: add API to open partition used for update
This adds function nxboot_open_update_partition that can be used as
a wrapper around nxboot_get_state call and simplifies the logic of
determining which partition is used as an update partition.

The function returns the opened file descriptor, therefore the user
does not need to figure out which partition he should open.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>